--- src/helpers.js
@@ -1,318 +1,0 @@
-import {isFinite, isArray, isObject, toFont, addRoundedRectPath, toTRBLCorners, valueOrDefault, drawPoint as drawPointOnChart} from 'chart.js/helpers';
-
-const widthCache = new Map();
-const toPercent = (s) => typeof s === 'string' && s.endsWith('%') && parseFloat(s) / 100;
-
-export function getSize(size, value) {
-  if (typeof value === 'number') {
-    return value;
-  } else if (typeof value === 'string') {
-    return toPercent(value) * size;
-  }
-  return size;
-}
-
-function calculateTextAlignment(size, options) {
-  const {x, width} = size;
-  const textAlign = options.textAlign;
-  if (textAlign === 'center') {
-    return x + width / 2;
-  } else if (textAlign === 'end' || textAlign === 'right') {
-    return x + width;
-  }
-  return x;
-}
-
-function getChartDimensionByScale(scale, options) {
-  if (scale) {
-    const min = scaleValue(scale, options.min, options.start);
-    const max = scaleValue(scale, options.max, options.end);
-    return {
-      start: Math.min(min, max),
-      end: Math.max(min, max)
-    };
-  }
-  return {
-    start: options.start,
-    end: options.end
-  };
-}
-
-/**
- * Apply border options to the canvas context before drawing a box
- * @param {CanvasRenderingContext2D} ctx - chart canvas context
- * @param {Object} options - options with border configuration
- * @returns {boolean} true is the border options have been applied
- */
-export function setBorderStyle(ctx, options) {
-  if (options && options.borderWidth) {
-    ctx.lineCap = options.borderCapStyle;
-    ctx.setLineDash(options.borderDash);
-    ctx.lineDashOffset = options.borderDashOffset;
-    ctx.lineJoin = options.borderJoinStyle;
-    ctx.lineWidth = options.borderWidth;
-    ctx.strokeStyle = options.borderColor;
-    return true;
-  }
-}
-
-function readValueToProps(value, props, defValue) {
-  const ret = {};
-  const objProps = isObject(props);
-  const keys = objProps ? Object.keys(props) : props;
-  const read = isObject(value)
-    ? objProps
-      ? prop => valueOrDefault(value[prop], value[props[prop]])
-      : prop => value[prop]
-    : () => value;
-
-  for (const prop of keys) {
-    ret[prop] = valueOrDefault(read(prop), defValue);
-  }
-  return ret;
-}
-
-export function toPosition(value) {
-  return readValueToProps(value, ['x', 'y'], 'center');
-}
-
-export const clamp = (x, from, to) => Math.min(to, Math.max(from, x));
-
-export function clampAll(obj, from, to) {
-  for (const key of Object.keys(obj)) {
-    obj[key] = clamp(obj[key], from, to);
-  }
-  return obj;
-}
-
-export function scaleValue(scale, value, fallback) {
-  value = typeof value === 'number' ? value : scale.parse(value);
-  return isFinite(value) ? scale.getPixelForValue(value) : fallback;
-}
-
-/**
- * Rotate a `point` relative to `center` point by `angle`
- * @param {{x: number, y: number}} point - the point to rotate
- * @param {{x: number, y: number}} center - center point for rotation
- * @param {number} angle - angle for rotation, in radians
- * @returns {{x: number, y: number}} rotated point
- */
-export function rotated(point, center, angle) {
-  var cos = Math.cos(angle);
-  var sin = Math.sin(angle);
-  var cx = center.x;
-  var cy = center.y;
-
-  return {
-    x: cx + cos * (point.x - cx) - sin * (point.y - cy),
-    y: cy + sin * (point.x - cx) + cos * (point.y - cy)
-  };
-}
-
-function isImageOrCanvas(content) {
-  return content instanceof Image || content instanceof HTMLCanvasElement;
-}
-
-/**
- * Measure the label size using the label options.
- * @param {CanvasRenderingContext2D} ctx - chart canvas context
- * @param {Object} options - options to configure the label
- * @returns {{width: number, height: number}} the measured size of the label
- */
-export function measureLabelSize(ctx, options) {
-  const content = options.content;
-  if (isImageOrCanvas(content)) {
-    return {
-      width: getSize(content.width, options.width),
-      height: getSize(content.height, options.height)
-    };
-  }
-  const font = toFont(options.font);
-  const lines = isArray(content) ? content : [content];
-  const mapKey = lines.join() + font.string;
-  if (!widthCache.has(mapKey)) {
-    ctx.save();
-    ctx.font = font.string;
-    const count = lines.length;
-    let width = 0;
-    for (let i = 0; i < count; i++) {
-      const text = lines[i];
-      width = Math.max(width, ctx.measureText(text).width);
-    }
-    ctx.restore();
-    const height = count * font.lineHeight;
-    widthCache.set(mapKey, {width, height});
-  }
-  return widthCache.get(mapKey);
-}
-
-/**
- * Draw a box with the size and the styling options.
- * @param {CanvasRenderingContext2D} ctx - chart canvas context
- * @param {{x: number, y: number, width: number, height: number}} rect - rect to draw
- * @param {Object} options - options to style the box
- * @returns {undefined}
- */
-export function drawBox(ctx, rect, options) {
-  const {x, y, width, height} = rect;
-  ctx.save();
-  const stroke = setBorderStyle(ctx, options);
-  ctx.fillStyle = options.backgroundColor;
-  ctx.beginPath();
-  addRoundedRectPath(ctx, {
-    x, y, w: width, h: height,
-    // TODO: v2 remove support for cornerRadius
-    radius: clampAll(toTRBLCorners(valueOrDefault(options.cornerRadius, options.borderRadius)), 0, Math.min(width, height) / 2)
-  });
-  ctx.closePath();
-  ctx.fill();
-  if (stroke) {
-    ctx.stroke();
-  }
-  ctx.restore();
-}
-
-const isEnabled = (options) => options && (options.display || options.enabled);
-
-export function isLabelVisible(options) {
-  return isEnabled(options) && options.content;
-}
-
-export function isPointVisible(options) {
-  return isEnabled(options) && options.radius > 0.1;
-}
-
-export function drawLabel(ctx, rect, options) {
-  if (!isLabelVisible(options)) {
-    return;
-  }
-  const content = options.content;
-  if (isImageOrCanvas(content)) {
-    ctx.drawImage(content, rect.x, rect.y, rect.width, rect.height);
-    return;
-  }
-  const labels = isArray(content) ? content : [content];
-  const font = toFont(options.font);
-  const lh = font.lineHeight;
-  const x = calculateTextAlignment(rect, options);
-  const y = rect.y + (lh / 2);
-  ctx.font = font.string;
-  ctx.textBaseline = 'middle';
-  ctx.textAlign = options.textAlign;
-  ctx.fillStyle = options.color;
-  labels.forEach((l, i) => ctx.fillText(l, x, y + (i * lh)));
-}
-
-export function getRectCenterPoint(rect) {
-  const {x, y, width, height} = rect;
-  return {
-    x: x + width / 2,
-    y: y + height / 2
-  };
-}
-
-export function getChartPoint(chart, options) {
-  const {chartArea, scales} = chart;
-  const xScale = scales[options.xScaleID];
-  const yScale = scales[options.yScaleID];
-  let x = chartArea.width / 2;
-  let y = chartArea.height / 2;
-
-  if (xScale) {
-    x = scaleValue(xScale, options.xValue, x);
-  }
-
-  if (yScale) {
-    y = scaleValue(yScale, options.yValue, y);
-  }
-  return {x, y};
-}
-
-export function getChartRect(chart, options) {
-  const xScale = chart.scales[options.xScaleID];
-  const yScale = chart.scales[options.yScaleID];
-  let {top: y, left: x, bottom: y2, right: x2} = chart.chartArea;
-
-  if (!xScale && !yScale) {
-    return {options: {}};
-  }
-
-  const xDim = getChartDimensionByScale(xScale, {min: options.xMin, max: options.xMax, start: x, end: x2});
-  x = xDim.start;
-  x2 = xDim.end;
-  const yDim = getChartDimensionByScale(yScale, {min: options.yMin, max: options.yMax, start: y, end: y2});
-  y = yDim.start;
-  y2 = yDim.end;
-
-  return {
-    x,
-    y,
-    x2,
-    y2,
-    width: x2 - x,
-    height: y2 - y
-  };
-}
-
-export function drawPoint(ctx, point, options) {
-  if (!isPointVisible(options)) {
-    return;
-  }
-  ctx.save();
-  ctx.fillStyle = options.backgroundColor;
-  setBorderStyle(ctx, options);
-  drawPointOnChart(ctx, options, point.x, point.y);
-  ctx.restore();
-}
-
-export function inPointRange(point, center, radius) {
-  if (!point || !center || radius <= 0) {
-    return false;
-  }
-  return (Math.pow(point.x - center.x, 2) + Math.pow(point.y - center.y, 2)) <= Math.pow(radius, 2);
-}
-
-export function inBoxRange(mouseX, mouseY, {x, y, width, height}) {
-  return mouseX >= x &&
-         mouseX <= x + width &&
-         mouseY >= y &&
-         mouseY <= y + height;
-}
-
-export function getElementCenterPoint(element, useFinalPosition) {
-  const {x, y} = element.getProps(['x', 'y'], useFinalPosition);
-  return {x, y};
-}
-
-export function getChartCircle(chart, options) {
-  const point = getChartPoint(chart, options);
-  return {
-    x: point.x,
-    y: point.y,
-    width: options.radius * 2,
-    height: options.radius * 2
-  };
-}
-
-export function isBoundToPoint(options) {
-  return options && (options.xValue || options.yValue);
-}
-
-export function resolvePointPosition(chart, options) {
-  if (!isBoundToPoint(options)) {
-    const box = getChartRect(chart, options);
-    const point = getRectCenterPoint(box);
-    let radius = options.radius;
-    if (!radius || isNaN(radius)) {
-      radius = Math.min(box.width, box.height) / 2;
-      options.radius = radius;
-    }
-    return {
-      x: point.x,
-      y: point.y,
-      width: radius * 2,
-      height: radius * 2
-    };
-  }
-  return getChartCircle(chart, options);
-}
